# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Live API
  version: 1.0.0
  description: |
    REST API for the Open Live broadcast production platform.

    **Authentication**: Pass an API key via `Authorization: Bearer <key>` header or `?key=<key>` query parameter.
    The following paths are exempt from authentication: `/health`, `/ready`, `/api/v1/status`, `/api/v1/server-info`, `/api/v1/reconnect`.
servers:
  - url: /
security:
  - bearerAuth: []
  - apiKeyQuery: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyQuery:
      type: apiKey
      in: query
      name: key
...
paths:
...
  /api/v1/whep-proxy:
    post:
      summary: WHEP offer proxy
      description: Proxies a WHEP SDP offer to the target Strom WHEP endpoint.
      operationId: whepProxyOffer
      tags: [WebRTC]
      parameters:
        - name: target
          in: query
          required: true
          schema:
            type: string
          description: URL-encoded target WHEP endpoint URL
      requestBody:
        required: true
        content:
          application/sdp:
            schema:
              type: string
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: SDP answer
          headers:
            Location:
              schema:
                type: string
          content:
            application/sdp:
              schema:
                type: string
        '502':
          description: Upstream error
    delete:
      summary: WHEP session teardown proxy
      operationId: whepProxyDelete
      tags: [WebRTC]
      parameters:
        - name: target
          in: query
          required: true
          schema:
            type: string
          description: URL-encoded target WHEP session URL
      responses:
        '204':
          description: Session closed

  # ── WHIP ────────────────────────────────────────────────────────────────────
  /api/v1/productions/{id}/whip/{mixerInput}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: mixerInput
        in: path
        required: true
        schema:
          type: string
        description: e.g. `video_in_1`
    post:
      summary: WHIP offer (start ingest)
      description: Initiates a WebRTC ingest session for the given mixer input slot.
      operationId: whipOffer
      tags: [WebRTC]
      requestBody:
        required: true
        content:
          application/sdp:
            schema:
              type: string
      responses:
        '201':
          description: SDP answer
          headers:
            Location:
              schema:
                type: string
          content:
            application/sdp:
              schema:
                type: string
    patch:
      summary: WHIP ICE trickle
      operationId: whipIceTrickle
      tags: [WebRTC]
      parameters:
        - name: session
          in: query
          schema:
            type: string
          description: URL-encoded session URL
      requestBody:
        required: true
        content:
          application/trickle-ice-sdpfrag:
            schema:
              type: string
      responses:
        '200':
          description: ICE candidate accepted
    delete:
      summary: WHIP session teardown
      operationId: whipDelete
      tags: [WebRTC]
      parameters:
        - name: session
          in: query
          schema:
            type: string
          description: URL-encoded session URL
      responses:
        '204':
          description: Session closed